+++ api/src/main/java/ru/geekbrains/store/api/OrderItemDto.java
@@ -1,13 +1,15 @@
 package ru.geekbrains.store.api;
 
 
+import java.math.BigDecimal;
+
 public class OrderItemDto {
 
     private Long id;
     private String productTitle;
     private int quantity;
-    private int pricePerProduct;
-    private int price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
 
     public Long getId() {
         return id;
@@ -33,23 +35,23 @@ public class OrderItemDto {
         this.quantity = quantity;
     }
 
-    public int getPricePerProduct() {
+    public BigDecimal getPricePerProduct() {
         return pricePerProduct;
     }
 
-    public void setPricePerProduct(int pricePerProduct) {
+    public void setPricePerProduct(BigDecimal pricePerProduct) {
         this.pricePerProduct = pricePerProduct;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
-    public OrderItemDto(Long id, String productTitle, int quantity, int pricePerProduct, int price) {
+    public OrderItemDto(Long id, String productTitle, int quantity, BigDecimal pricePerProduct, BigDecimal price) {
         this.id = id;
         this.productTitle = productTitle;
         this.quantity = quantity;
