+++ api/src/main/java/ru/geekbrains/store/api/ProductDto.java
@@ -1,13 +1,15 @@
 package ru.geekbrains.store.api;
 
 
+import java.math.BigDecimal;
+
 public class ProductDto {
 
     private Long id;
     private String title;
-    private int price;
+    private BigDecimal price;
 
-    public ProductDto(Long id, String title, int price) {
+    public ProductDto(Long id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -32,11 +34,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
